Put FontInitData into separate file - add \"is_fixed\" edge property
