+++ src/algorithm/Domain.h
@@ -446,6 +446,7 @@ public:
   Edge& add_fixed_edge( Vertex& v1, Vertex& v2 )
   {
     Edge& new_edge = fixed_edges_.add_edge(v1, v2, INTERIOR_EDGE_COLOR);
+    new_edge.add_property( EdgeProperty::is_fixed );
     return new_edge;
 
   } // Domain::add_fixed_edge()
